+++ SearchThing/src/Fusion/FusionSpawnHistory.cs
@@ -1,34 +1,43 @@
-﻿using Il2CppSLZ.Marrow.Warehouse;
+﻿using HarmonyLib;
+using Il2CppSLZ.Marrow.Warehouse;
+using LabFusion.Data;
 using LabFusion.Player;
+using MelonLoader;
 using SearchThing.History;
 using SearchThing.Search;
 using SearchThing.Util;
 
 namespace SearchThing.Fusion;
 
-public record FusionSpawnHistoryEntry(ISearchableCrate Crate, PlayerID SpawnerId) : IHistorySearchableCrate
+public record FusionSpawnHistoryEntry(Barcode Barcode, PlayerID? SpawnerId) : IHistorySearchableCrate
 {
-    public SearchTag Name => Crate.Name;
-    public SearchTag PalletName => Crate.PalletName;
-    public SearchTag Author => Crate.Author;
-    public SearchTag[] Tags => Crate.Tags;
-    public string Description => Crate.Description;
-    public bool Redacted => Crate.Redacted;
-    public CrateType CrateType => Crate.CrateType;
-    public CrateSubType CrateSubType => Crate.CrateSubType;
-    public int Salt => Crate.Salt;
-    public int Score => Crate.Score;
+    private static readonly SearchTag UnknownCrateTag = new SearchTag("Unknown Crate");
+    public ISearchableCrate? Crate { get; } = GlobalCrateManager.GetCrate(Barcode); 
+    
+    public SearchTag Name => Crate?.Name ?? UnknownCrateTag;
+    public SearchTag PalletName => Crate?.PalletName ?? UnknownCrateTag;
+    public SearchTag Author => Crate?.Author ?? UnknownCrateTag;
+    public SearchTag[] Tags => Crate?.Tags ?? Array.Empty<SearchTag>();
+    public string Description => Crate?.Description ?? "No description available.";
+    public bool Redacted => Crate?.Redacted ?? false;
+    public CrateType CrateType => Crate?.CrateType ?? CrateType.Invalid;
+    public CrateSubType CrateSubType => Crate?.CrateSubType ?? CrateSubType.None;
+    public int Salt => Crate?.Salt ?? 0;
+    public int Score => Crate?.Score ?? 0;
     public DateTime DateAdded { get; set; }
-    public Barcode Barcode => Crate.Barcode;
 }
 
 public static class FusionSpawnHistory
 {
     private static readonly BoundCrateList<FusionSpawnHistoryEntry> BoundEntries = new();
     
-    public static void AddEntry(Crate crate, PlayerID spawnerId)
+    public static void AddEntry(string barcode, byte? smallPlayerId)
     {
-        BoundEntries.AddCrate(new FusionSpawnHistoryEntry(new SearchableCrate(crate), spawnerId));
+        var spawnerId = smallPlayerId.HasValue ? PlayerIDManager.GetPlayerID(smallPlayerId.Value) : null;
+        
+        var barcodeObj = new Barcode(barcode);
+        
+        BoundEntries.AddCrate(new FusionSpawnHistoryEntry(barcodeObj, spawnerId));
     }
     
     public static void SearchAsync(string query, ISearchOrder order, Func<ISearchableCrate, bool> filter, Action<SearchResults<FusionSpawnHistoryEntry>> callback)