+++ SearchThing/src/Fusion/FusionSpawnHistoryPage.cs
@@ -1,5 +1,7 @@
 ﻿using Il2CppSLZ.Marrow.Warehouse;
+using LabFusion.Data;
 using LabFusion.Network;
+using LabFusion.Safety;
 using SearchThing.Extensions;
 using SearchThing.Extensions.Pages;
 using SearchThing.Extensions.Panel;
@@ -14,11 +16,43 @@ namespace SearchThing.Fusion;
 
 public class FusionSpawnHistoryPage : BasicSearchPanel<FusionSpawnHistoryEntry>
 {
-    public override string Tag { get; } = "Fusion Spawn History";
+    private static readonly Sprite BlockIcon = ImageHelper.LoadEmbeddedSprite("SearchThing.resources.BlockIcon.png");
+    
+    public override string Tag => "Fusion Spawn History";
+    public override bool IsVisible => NetworkInfo.HasServer;
+    public override bool CanAssign => false;
+
+    public override bool ResearchOnPageChange => true;
+    public override Sprite ItemFunctionIcon => BlockIcon;
+
+    public override bool HasItemFunction => true;
+
+    public override Color? GetItemFunctionHighlight(SpawnablePanelExtension extension, ISearchableCrate? crate)
+    {
+        return crate != null && FusionBlacklistHelper.IsBlacklisted(crate.Barcode._id) ? Color.red : Color.green;
+    }
+
+    public override void OnItemFunction(SpawnablePanelExtension extension, ISearchableCrate crate)
+    {
+        FusionBlacklistHelper.ToggleBlacklist(crate.Barcode._id);
+        
+        extension.RequestRefresh();
+    }
+    
+    // Change search order
+    public override ISelectableSearchOrder[] SupportedOrders { get; } = {
+        new DateNewAddedSearchOrder(),
+        new DateOldAddedSearchOrder(),
+        new ScoreSearchOrder(),
+        new AlphabeticalSearchOrder(),
+        new RandomSearchOrder()
+    };
 
     public override ItemRenderData GetRenderDataForCrate(FusionSpawnHistoryEntry crate)
     {
-        var ownerName = crate.SpawnerId.TryGetDisplayName(out var name) ? name : "Unknown";
+        var ownerName = crate.SpawnerId?.IsValid == true && crate.SpawnerId.TryGetDisplayName(out var name) 
+            ? StringHelper.RemoveUnityRichText(name) 
+            : "Unknown";
         
         return new ItemRenderData
         {
@@ -33,6 +67,6 @@ public class FusionSpawnHistoryPage : BasicSearchPanel<FusionSpawnHistoryEntry>
 
     protected override void Search(string query, ISearchOrder order, Action<SearchResults<FusionSpawnHistoryEntry>> callback)
     {
-        FusionSpawnHistory.SearchAsync(query, order, _ => true, callback);
+        FusionSpawnHistory.SearchAsync(query, order, c => c.CrateType == CrateType.Prop, callback);
     }
 }