+++ SearchThing/src/Mod.cs
@@ -1,11 +1,15 @@
 ﻿using BoneLib;
+using LabFusion.Utilities;
 using SearchThing;
 using MelonLoader;
+using SearchThing.Fusion;
+using SearchThing.Patches;
+using SearchThing.Patches.Compatibility;
 using SearchThing.Presets;
 using SearchThing.Search;
 using UnityEngine;
 
-[assembly: MelonInfo(typeof(Mod), "SearchThing", "0.4.3", "Mash")]
+[assembly: MelonInfo(typeof(Mod), "SearchThing", "0.5.0", "Mash")]
 [assembly: MelonGame("Stress Level Zero", "BONELAB")]
 namespace SearchThing;
 
@@ -19,11 +23,26 @@ public class Mod : MelonMod
         
         Hooking.OnWarehouseReady += OnWarehouseReady;
     }
+
+    /// <summary>
+    /// Separate method to prevent loading errors if Fusion is not present.
+    /// </summary>
+    private static void HookFusion()
+    {
+        FusionNetworkSpawnPatch.TryInitialize();
+        MultiplayerHooking.OnTargetLevelLoaded += SpawnGunPatches.ClearSelectedCrate;
+    }
     
     private static void OnWarehouseReady()
     {
         SearchManager.InitializeSearchThread();
         PresetManager.LoadPresets();
+        
+        if (IsFusionLoaded)
+        {
+            HookFusion();
+        }
+        
     }
 
     public override void OnDeinitializeMelon()