+++ SearchThing/src/Util/SearchTag.cs
@@ -27,17 +27,14 @@ public struct SearchTag
     
     // Preperation
     
-    private static readonly Regex UnityRichTextRegex = new(@"</?[a-zA-Z]*=[^>]*>|</?[a-zA-Z]+>", RegexOptions.Compiled);
-    private static readonly Regex NonAlphanumericRegex = new(@"[^ a-zA-Z0-9]", RegexOptions.Compiled);
-    
     public static string Preprocess(string value)
     {
         // To lower case
         var lower = value.ToLowerInvariant();
         // Remove non-alphanumeric characters
-        var nonAlphanumericStripped = NonAlphanumericRegex.Replace(lower, "");
+        var nonAlphanumericStripped = StringHelper.RemoveNonAlphanumeric(lower);
         // Remove Unity rich text tags
-        var nonUnityRichTextStripped = UnityRichTextRegex.Replace(nonAlphanumericStripped, "");
+        var nonUnityRichTextStripped = StringHelper.RemoveUnityRichText(nonAlphanumericStripped);
         // Remove spaces
         var spaceRemoved = nonUnityRichTextStripped.Replace(" ", "");
         // Trim leading and trailing whitespace